% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD.R
\name{IBD}
\alias{IBD}
\title{Title}
\usage{
IBD(
  parents,
  coord,
  coord_cols = c("longitude", "latitude"),
  selfing = FALSE,
  limit = -1
)
}
\arguments{
\item{parents}{data.table}

\item{coord}{data.table}

\item{coord_cols}{character vector}

\item{selfing}{logical}

\item{limit}{numeric}
}
\value{
data.table
}
\description{
Title
}
