% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_mating.R
\name{mixed_mating}
\alias{mixed_mating}
\title{Title}
\usage{
mixed_mating(parents, inbreeding_index = 0.3)
}
\arguments{
\item{parents}{data.table}

\item{inbreeding_index}{numeric}
}
\value{
data.table
}
\description{
Title
}
