% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offspring-package.R
\docType{package}
\name{offspring-package}
\alias{offspring}
\alias{offspring-package}
\title{offspring: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Nikos Pechlivanis \email{npechlv@gmail.com} (\href{https://orcid.org/0000-0003-2502-612X}{ORCID})

Authors:
\itemize{
  \item Fotis Psomopoulos \email{fpsom@certh.gr} (\href{https://orcid.org/0000-0002-0222-4273}{ORCID})
  \item Aristotelis Papageorgiou \email{apapage@mbg.duth.gr} (\href{https://orcid.org/0000-0001-6657-7820}{ORCID})
}

}
\keyword{internal}
