% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{population}
\alias{population}
\title{Title}
\usage{
population(parents, structure, complete = TRUE, gene_flow_rate = 0.05)
}
\arguments{
\item{parents}{data.table}

\item{structure}{data.table}

\item{complete}{logical}

\item{gene_flow_rate}{numeric}
}
\value{
data.table
}
\description{
Title
}
