% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_mating.R
\name{random_mating}
\alias{random_mating}
\title{Title}
\usage{
random_mating(parents, mating = "random", selfing = TRUE)
}
\arguments{
\item{parents}{data.table}

\item{mating}{character}

\item{selfing}{logical}
}
\value{
data.table
}
\description{
Title
}
